% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_creation.R
\name{create_select_all_query}
\alias{create_select_all_query}
\title{Create a simple query for Metabase}
\usage{
create_select_all_query(table_id, database_id)
}
\arguments{
\item{table_id}{ID of the table to be used.}

\item{database_id}{ID of the database to be used.}
}
\value{
A R representation of the query json.
}
\description{
Create a simple query for Metabase
}
