% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_creation.R
\name{create_select_specific_columns}
\alias{create_select_specific_columns}
\title{Create a query for Metabase with specific columns}
\usage{
create_select_specific_columns(table_id, database_id, columns)
}
\arguments{
\item{table_id}{ID of the table to be used.}

\item{database_id}{ID of the database to be used.}

\item{columns}{vector of column ids to select}
}
\value{
A R representation of the query json.
}
\description{
Create a query for Metabase with specific columns
}
