% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_creation.R
\name{unique_id_lookup}
\alias{unique_id_lookup}
\title{Id look up with uniqueness validation}
\usage{
unique_id_lookup(data, search_name, name_col, id_col)
}
\arguments{
\item{data}{Dataframe for id look up.}

\item{search_name}{Value for the look up}

\item{name_col}{Column in which to search for search_name}

\item{id_col}{Column in which the id is found}
}
\value{
The id value corresponding to search_name in column name_col
}
\description{
Id look up with uniqueness validation
}
